+++ src/components/dialogs/StudentDataManagementDialog.vue
@@ -73,7 +73,7 @@ import ContentApiService from '@/services/contentApiService';
 import { useStore } from 'vuex';
 import { useI18n } from 'vue-i18n';
 import EmployeeDataForm from '@/components/forms/EmployeeDataForm.vue';
-import FilesUploadForm from '@/components/forms/FilesUploadForm.vue';
+import FilesUploadForm from '@/components/forms/StudentFilesUploadForm.vue';
 import CustomDialog from '@/components/dialogs/base/CustomDialog.vue';
 
 const props = defineProps({
